✨ (Skills) - Add hover animation on skills + move BentoCard component to separate files
