+++ .eslintrc.json
@@ -1,6 +1,3 @@
 {
-  "extends": ["next/core-web-vitals", "plugin:tailwindcss/recommended"],
-  "rules": {
-    "tailwindcss/no-custom-classname": "off"
-  }
+  "extends": ["next/core-web-vitals", "plugin:tailwindcss/recommended"]
 }
